created task policy

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\TaskResource;
 use App\Models\Task;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -15,6 +16,7 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
@@ -23,6 +25,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if(request()->user()->cannot('create', Task::class)) {
+            abort(403, "This action is unauthorized.");
+        }
+
         $task = $request->user()->tasks()->create($request->validated());
         return TaskResource::make($task);
     }
@@ -32,6 +38,7 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
         return TaskResource::make($task);
     }
 
@@ -40,6 +47,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if($request->user()->cannot('update', $task)) {
+            abort(403, "This action is unauthorized.");
+        }
+
         $task->update($request->validated());
         return TaskResource::make($task);
     }
@@ -49,6 +60,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if(request()->user()->cannot('delete', $task)) {
+            abort(403, "This action is unauthorized.");
+        }
+
         $task->delete();
         return response()->noContent();
     }
